Разделяет на модули файл main
