+++ js/helpers/creater.js
@@ -1,0 +1,3 @@
+export const createElements = (creater, count) => (
+  Array.from({ length: count }, creater)
+);
