+++ js/helpers/id.js
@@ -1,0 +1,4 @@
+export const getIdGenerator = () => {
+  let id = 1;
+  return () => id++;
+};
