+++ js/mock/descriptions.js
@@ -1,0 +1,8 @@
+export const DESCRIPTIONS = [
+  'Крутяк',
+  'Прикольно',
+  'Необычно',
+  'Невероятно',
+  'Необыкновенно',
+  'Оригинально',
+];
