+++ js/mock/names.js
@@ -1,0 +1,8 @@
+export const NAMES = [
+  'Вася',
+  'Саша',
+  'Никита',
+  'Валера',
+  'Гоша',
+  'Жора',
+];
